w3c-reports/12092014/gtk/runned-tests/html/semantics/forms/constraints/form-validation-willValidate-expected.txt



PASS [INPUT in HIDDEN status] Must be barred from the constraint validation 
PASS [INPUT in BUTTON status] Must be barred from the constraint validation 
PASS [INPUT in RESET status] Must be barred from the constraint validation 
PASS [BUTTON in BUTTON status] Must be barred from the constraint validation 
PASS [BUTTON in RESET status] Must be barred from the constraint validation 
FAIL [fieldset]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:162:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_willValidate@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:153:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:321:27
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [output]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:162:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_willValidate@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:153:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:321:27
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
PASS [object]  Must be barred from the constraint validation 
PASS [keygen]  Must be barred from the constraint validation 
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in TEXT status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in TEXT status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in SEARCH status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in SEARCH status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in TEL status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in TEL status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in URL status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in URL status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in URL status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in URL status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in EMAIL status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in EMAIL status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in PASSWORD status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in PASSWORD status] The willValidate attribute must be false if it has a datalist ancestor 
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [INPUT in DATE status] The date type must be supported. assert_equals: The date type should be supported. expected "date" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [INPUT in MONTH status] The month type must be supported. assert_equals: The month type should be supported. expected "month" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [INPUT in WEEK status] The week type must be supported. assert_equals: The week type should be supported. expected "week" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [INPUT in TIME status] The time type must be supported. assert_equals: The time type should be supported. expected "time" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
FAIL [INPUT in COLOR status] The color type must be supported. assert_equals: The color type should be supported. expected "color" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/constraints/support/validator.js:214:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_support_type@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:213:9
run_test@http://localhost:8800/html/semantics/forms/constraints/support/validator.js:289:35
global code@http://localhost:8800/html/semantics/forms/constraints/form-validation-willValidate.html:89:21)
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in FILE status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in FILE status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is disabled 
PASS [INPUT in SUBMIT status] The willValidate attribute must be true if an elment is mutable 
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is readonly 
PASS [INPUT in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [BUTTON in SUBMIT status] Must be barred from the constraint validation 
PASS [BUTTON in SUBMIT status] The willValidate attribute must be true if an elment is mutable 
PASS [BUTTON in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor 
PASS [select]  Must be barred from the constraint validation 
PASS [select]  The willValidate attribute must be true if an elment is mutable 
PASS [select]  The willValidate attribute must be false if it has a datalist ancestor 
PASS [textarea]  Must be barred from the constraint validation 
PASS [textarea]  The willValidate attribute must be true if an elment is mutable 
PASS [textarea]  The willValidate attribute must be false if it has a datalist ancestor 
